+++ extension/src/contentFunctions/styling/bulk.ts
